split out client/server transports, re-enable network edge case tests
