+++ transport/transport.ts
@@ -4,15 +4,10 @@ import {
   OpaqueTransportMessage,
   OpaqueTransportMessageSchema,
   TransportClientId,
-  ControlMessageHandshakeRequestSchema,
-  ControlMessageHandshakeResponseSchema,
-  handshakeRequestMessage,
-  handshakeResponseMessage,
   PartialTransportMessage,
   ControlFlags,
   ControlMessagePayloadSchema,
   isAck,
-  PROTOCOL_VERSION,
 } from './message';
 import {
   BaseLogger,
@@ -28,22 +23,15 @@ import {
   ProtocolError,
   ProtocolErrorType,
 } from './events';
-import { Connection, Session, SessionOptions } from './session';
+import { Connection, Session } from './session';
 import { Static } from '@sinclair/typebox';
-import { coerceErrorString } from '../util/stringify';
-import { ConnectionRetryOptions, LeakyBucketRateLimit } from './rateLimit';
-import { NaiveJsonCodec } from '../codec';
-import tracer, {
-  PropagationContext,
-  createConnectionTelemetryInfo,
-  getPropagationContext,
-} from '../tracing';
+import { PropagationContext, createConnectionTelemetryInfo } from '../tracing';
 import { SpanStatusCode } from '@opentelemetry/api';
-import { ParsedMetadata } from '../router/context';
 import {
-  ClientHandshakeOptions,
-  ServerHandshakeOptions,
-} from '../router/handshake';
+  ProvidedTransportOptions,
+  TransportOptions,
+  defaultTransportOptions,
+} from './options';
 
 /**
  * Represents the possible states of a transport.
@@ -51,42 +39,6 @@ import {
  * @property {'closed'} closed - The transport is permanently closed and cannot be reopened.
  */
 export type TransportStatus = 'open' | 'closed';
-
-type TransportOptions = SessionOptions;
-
-export type ProvidedTransportOptions = Partial<TransportOptions>;
-
-export const defaultTransportOptions: TransportOptions = {
-  heartbeatIntervalMs: 1_000,
-  heartbeatsUntilDead: 2,
-  sessionDisconnectGraceMs: 5_000,
-  codec: NaiveJsonCodec,
-};
-
-type ClientTransportOptions = TransportOptions & ConnectionRetryOptions;
-
-export type ProvidedClientTransportOptions = Partial<ClientTransportOptions>;
-
-const defaultConnectionRetryOptions: ConnectionRetryOptions = {
-  baseIntervalMs: 250,
-  maxJitterMs: 200,
-  maxBackoffMs: 32_000,
-  attemptBudgetCapacity: 5,
-  budgetRestoreIntervalMs: 200,
-};
-
-const defaultClientTransportOptions: ClientTransportOptions = {
-  ...defaultTransportOptions,
-  ...defaultConnectionRetryOptions,
-};
-
-type ServerTransportOptions = TransportOptions;
-
-export type ProvidedServerTransportOptions = Partial<ServerTransportOptions>;
-
-const defaultServerTransportOptions: ServerTransportOptions = {
-  ...defaultTransportOptions,
-};
 
 /**
  * Transports manage the lifecycle (creation/deletion) of sessions and connections. Its responsibilities include:
@@ -548,756 +500,5 @@ export abstract class Transport<ConnType extends Connection> {
 
   getStatus(): TransportStatus {
     return this.status;
-  }
-}
-
-export abstract class ClientTransport<
-  ConnType extends Connection,
-> extends Transport<ConnType> {
-  /**
-   * The options for this transport.
-   */
-  protected options: ClientTransportOptions;
-
-  /**
-   * The map of reconnect promises for each client ID.
-   */
-  inflightConnectionPromises: Map<TransportClientId, Promise<ConnType>>;
-  retryBudget: LeakyBucketRateLimit;
-
-  /**
-   * A flag indicating whether the transport should automatically reconnect
-   * when a connection is dropped.
-   * Realistically, this should always be true for clients unless you are writing
-   * tests or a special case where you don't want to reconnect.
-   */
-  reconnectOnConnectionDrop = true;
-
-  /**
-   * Optional handshake options for this client.
-   */
-  handshakeExtensions?: ClientHandshakeOptions;
-
-  constructor(
-    clientId: TransportClientId,
-    providedOptions?: ProvidedClientTransportOptions,
-  ) {
-    super(clientId, providedOptions);
-    this.options = {
-      ...defaultClientTransportOptions,
-      ...providedOptions,
-    };
-    this.inflightConnectionPromises = new Map();
-    this.retryBudget = new LeakyBucketRateLimit(this.options);
-  }
-
-  extendHandshake(options: ClientHandshakeOptions) {
-    this.handshakeExtensions = options;
-  }
-
-  protected handleConnection(conn: ConnType, to: TransportClientId): void {
-    if (this.getStatus() !== 'open') return;
-    let session: Session<ConnType> | undefined = undefined;
-
-    // kill the conn after the grace period if we haven't received a handshake
-    const handshakeTimeout = setTimeout(() => {
-      if (session) return;
-      this.log?.warn(
-        `connection to ${to} timed out waiting for handshake, closing`,
-        { ...conn.loggingMetadata, clientId: this.clientId, connectedTo: to },
-      );
-      conn.close();
-    }, this.options.sessionDisconnectGraceMs);
-
-    const handshakeHandler = (data: Uint8Array) => {
-      const maybeSession = this.receiveHandshakeResponseMessage(data, conn);
-      clearTimeout(handshakeTimeout);
-      if (!maybeSession) {
-        conn.close();
-        return;
-      } else {
-        session = maybeSession;
-      }
-
-      // when we are done handshake sequence,
-      // remove handshake listener and use the normal message listener
-      conn.removeDataListener(handshakeHandler);
-      conn.addDataListener((data) => {
-        const parsed = this.parseMsg(data, conn);
-        if (!parsed) {
-          conn.telemetry?.span.setStatus({
-            code: SpanStatusCode.ERROR,
-            message: 'message parse failure',
-          });
-          conn.close();
-          return;
-        }
-
-        this.handleMsg(parsed, conn);
-      });
-    };
-
-    conn.addDataListener(handshakeHandler);
-    conn.addCloseListener(() => {
-      if (session) {
-        this.onDisconnect(conn, session);
-      }
-
-      const willReconnect =
-        this.reconnectOnConnectionDrop && this.getStatus() === 'open';
-
-      this.log?.info(
-        `connection to ${to} disconnected` +
-          (willReconnect ? ', reconnecting' : ''),
-        {
-          ...conn.loggingMetadata,
-          ...session?.loggingMetadata,
-          clientId: this.clientId,
-          connectedTo: to,
-        },
-      );
-
-      this.inflightConnectionPromises.delete(to);
-      if (this.reconnectOnConnectionDrop) {
-        void this.connect(to);
-      }
-    });
-    conn.addErrorListener((err) => {
-      conn.telemetry?.span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: 'connection error',
-      });
-      this.log?.warn(
-        `error in connection to ${to}: ${coerceErrorString(err)}`,
-        {
-          ...conn.loggingMetadata,
-          ...session?.loggingMetadata,
-          clientId: this.clientId,
-          connectedTo: to,
-        },
-      );
-    });
-  }
-
-  receiveHandshakeResponseMessage(
-    data: Uint8Array,
-    conn: ConnType,
-  ): Session<ConnType> | false {
-    const parsed = this.parseMsg(data, conn);
-    if (!parsed) {
-      conn.telemetry?.span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: 'non-transport message',
-      });
-      this.protocolError(
-        ProtocolError.HandshakeFailed,
-        'received non-transport message',
-      );
-      return false;
-    }
-
-    if (!Value.Check(ControlMessageHandshakeResponseSchema, parsed.payload)) {
-      conn.telemetry?.span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: 'invalid handshake response',
-      });
-      this.log?.warn(`received invalid handshake resp`, {
-        ...conn.loggingMetadata,
-        clientId: this.clientId,
-        connectedTo: parsed.from,
-        transportMessage: parsed,
-        validationErrors: [
-          ...Value.Errors(
-            ControlMessageHandshakeResponseSchema,
-            parsed.payload,
-          ),
-        ],
-      });
-      this.protocolError(
-        ProtocolError.HandshakeFailed,
-        'invalid handshake resp',
-      );
-      return false;
-    }
-
-    if (!parsed.payload.status.ok) {
-      conn.telemetry?.span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: 'handshake rejected',
-      });
-      this.log?.warn(`received handshake rejection`, {
-        ...conn.loggingMetadata,
-        clientId: this.clientId,
-        connectedTo: parsed.from,
-        transportMessage: parsed,
-      });
-      this.protocolError(
-        ProtocolError.HandshakeFailed,
-        parsed.payload.status.reason,
-      );
-      return false;
-    }
-
-    // before we claim victory and we deem that the handshake is fully established, check that our
-    // session matches the remote's. if they do not match, proactively close the connection.
-    // otherwise we will end up breaking a lot of invariants.
-    //
-    // note that there may be a way to convey the expectation of the client that this is a
-    // transparent reconnect to the server so that the server can do the same in a single roundtrip,
-    // but that requires a bit more surgery to the handshake.
-    const previousSession = this.sessions.get(parsed.from);
-    if (
-      previousSession?.advertisedSessionId &&
-      previousSession.advertisedSessionId !== parsed.payload.status.sessionId
-    ) {
-      this.deleteSession({
-        session: previousSession,
-        closeHandshakingConnection: true,
-      });
-
-      conn.telemetry?.span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: 'session id mismatch',
-      });
-      this.log?.warn(`handshake from ${parsed.from} session id mismatch`, {
-        ...conn.loggingMetadata,
-        clientId: this.clientId,
-        connectedTo: parsed.from,
-        transportMessage: parsed,
-      });
-      this.protocolError(ProtocolError.HandshakeFailed, 'session id mismatch');
-      return false;
-    }
-
-    this.log?.debug(`handshake from ${parsed.from} ok`, {
-      ...conn.loggingMetadata,
-      clientId: this.clientId,
-      connectedTo: parsed.from,
-      transportMessage: parsed,
-    });
-
-    const { session, isTransparentReconnect } = this.getOrCreateSession({
-      to: parsed.from,
-      conn,
-      sessionId: parsed.payload.status.sessionId,
-    });
-
-    this.onConnect(conn, session, isTransparentReconnect);
-
-    // After a successful connection, we start restoring the budget
-    // so that the next time we try to connect, we don't hit the client
-    // with backoff forever.
-    this.retryBudget.startRestoringBudget(session.to);
-    return session;
-  }
-
-  /**
-   * Abstract method that creates a new {@link Connection} object.
-   * This should call {@link handleConnection} when the connection is created.
-   * The downstream client implementation needs to implement this.
-   *
-   * @param to The client ID of the node to connect to.
-   * @returns The new connection object.
-   */
-  protected abstract createNewOutgoingConnection(
-    to: TransportClientId,
-  ): Promise<ConnType>;
-
-  /**
-   * Manually attempts to connect to a client.
-   * @param to The client ID of the node to connect to.
-   */
-  async connect(to: TransportClientId): Promise<void> {
-    if (this.connections.has(to)) {
-      this.log?.info(`already connected to ${to}, skipping connect attempt`, {
-        clientId: this.clientId,
-        connectedTo: to,
-      });
-      return;
-    }
-
-    const canProceedWithConnection = () => this.getStatus() === 'open';
-    if (!canProceedWithConnection()) {
-      this.log?.info(
-        `transport state is no longer open, cancelling attempt to connect to ${to}`,
-        { clientId: this.clientId, connectedTo: to },
-      );
-      return;
-    }
-
-    let reconnectPromise = this.inflightConnectionPromises.get(to);
-    if (!reconnectPromise) {
-      // check budget
-      if (!this.retryBudget.hasBudget(to)) {
-        const budgetConsumed = this.retryBudget.getBudgetConsumed(to);
-        const errMsg = `tried to connect to ${to} but retry budget exceeded (more than ${budgetConsumed} attempts in the last ${this.retryBudget.totalBudgetRestoreTime}ms)`;
-        this.log?.error(errMsg, { clientId: this.clientId, connectedTo: to });
-        this.protocolError(ProtocolError.RetriesExceeded, errMsg);
-        return;
-      }
-
-      let sleep = Promise.resolve();
-      const backoffMs = this.retryBudget.getBackoffMs(to);
-      if (backoffMs > 0) {
-        sleep = new Promise((resolve) => setTimeout(resolve, backoffMs));
-      }
-
-      this.log?.info(
-        `attempting connection to ${to} (${backoffMs}ms backoff)`,
-        {
-          clientId: this.clientId,
-          connectedTo: to,
-        },
-      );
-      this.retryBudget.consumeBudget(to);
-      reconnectPromise = tracer.startActiveSpan('connect', async (span) => {
-        try {
-          span.addEvent('backoff', { backoffMs });
-          await sleep;
-          if (!canProceedWithConnection()) {
-            throw new Error('transport state is no longer open');
-          }
-
-          span.addEvent('connecting');
-          const conn = await this.createNewOutgoingConnection(to);
-          if (!canProceedWithConnection()) {
-            this.log?.info(
-              `transport state is no longer open, closing pre-handshake connection to ${to}`,
-              {
-                ...conn.loggingMetadata,
-                clientId: this.clientId,
-                connectedTo: to,
-              },
-            );
-            conn.close();
-            throw new Error('transport state is no longer open');
-          }
-
-          span.addEvent('sending handshake');
-          const ok = await this.sendHandshake(to, conn);
-          if (!ok) {
-            conn.close();
-            throw new Error('failed to send handshake');
-          }
-
-          return conn;
-        } catch (err) {
-          // rethrow the error so that the promise is rejected
-          // as it was before we wrapped it in a span
-          const errStr = coerceErrorString(err);
-          span.recordException(errStr);
-          span.setStatus({ code: SpanStatusCode.ERROR });
-          throw err;
-        } finally {
-          span.end();
-        }
-      });
-
-      this.inflightConnectionPromises.set(to, reconnectPromise);
-    } else {
-      this.log?.info(
-        `attempting connection to ${to} (reusing previous attempt)`,
-        {
-          clientId: this.clientId,
-          connectedTo: to,
-        },
-      );
-    }
-
-    try {
-      await reconnectPromise;
-    } catch (error: unknown) {
-      this.inflightConnectionPromises.delete(to);
-      const errStr = coerceErrorString(error);
-
-      if (!this.reconnectOnConnectionDrop || !canProceedWithConnection()) {
-        this.log?.warn(`connection to ${to} failed (${errStr})`, {
-          clientId: this.clientId,
-          connectedTo: to,
-        });
-      } else {
-        this.log?.warn(`connection to ${to} failed (${errStr}), retrying`, {
-          clientId: this.clientId,
-          connectedTo: to,
-        });
-        await this.connect(to);
-      }
-    }
-  }
-
-  protected deleteSession({
-    session,
-    closeHandshakingConnection,
-    handshakingConn,
-  }: {
-    session: Session<ConnType>;
-    closeHandshakingConnection: boolean;
-    handshakingConn?: ConnType;
-  }) {
-    this.inflightConnectionPromises.delete(session.to);
-    super.deleteSession({
-      session,
-      closeHandshakingConnection,
-      handshakingConn,
-    });
-  }
-
-  protected async sendHandshake(to: TransportClientId, conn: ConnType) {
-    let metadata: unknown = undefined;
-
-    if (this.handshakeExtensions) {
-      metadata = await this.handshakeExtensions.construct();
-      if (!Value.Check(this.handshakeExtensions.schema, metadata)) {
-        this.log?.error(`constructed handshake metadata did not match schema`, {
-          ...conn.loggingMetadata,
-          clientId: this.clientId,
-          connectedTo: to,
-          validationErrors: [
-            ...Value.Errors(this.handshakeExtensions.schema, metadata),
-          ],
-          tags: ['invariant-violation'],
-        });
-        this.protocolError(
-          ProtocolError.HandshakeFailed,
-          'handshake metadata did not match schema',
-        );
-        conn.telemetry?.span.setStatus({
-          code: SpanStatusCode.ERROR,
-          message: 'handshake meta mismatch',
-        });
-        return false;
-      }
-    }
-
-    // dont pass conn here as we dont want the session to start using the conn
-    // until we have finished the handshake. Still, let the session know that
-    // it is semi-associated with the conn, and it can close it if .close() is called.
-    const { session } = this.getOrCreateSession({ to, handshakingConn: conn });
-    const requestMsg = handshakeRequestMessage(
-      this.clientId,
-      to,
-      session.id,
-      metadata,
-      getPropagationContext(session.telemetry.ctx),
-    );
-    this.log?.debug(`sending handshake request to ${to}`, {
-      ...conn.loggingMetadata,
-      clientId: this.clientId,
-      connectedTo: to,
-      transportMessage: requestMsg,
-    });
-    conn.send(this.codec.toBuffer(requestMsg));
-    return true;
-  }
-
-  close() {
-    this.retryBudget.close();
-    super.close();
-  }
-}
-
-export abstract class ServerTransport<
-  ConnType extends Connection,
-> extends Transport<ConnType> {
-  /**
-   * The options for this transport.
-   */
-  protected options: ServerTransportOptions;
-
-  /**
-   * Optional handshake options for the server.
-   */
-  handshakeExtensions?: ServerHandshakeOptions;
-
-  /**
-   * A map of session handshake data for each session.
-   */
-  sessionHandshakeMetadata: WeakMap<Session<ConnType>, ParsedMetadata>;
-
-  constructor(
-    clientId: TransportClientId,
-    providedOptions?: ProvidedServerTransportOptions,
-  ) {
-    super(clientId, providedOptions);
-    this.options = {
-      ...defaultServerTransportOptions,
-      ...providedOptions,
-    };
-    this.sessionHandshakeMetadata = new WeakMap();
-    this.log?.info(`initiated server transport`, {
-      clientId: this.clientId,
-      protocolVersion: PROTOCOL_VERSION,
-    });
-  }
-
-  extendHandshake(options: ServerHandshakeOptions) {
-    this.handshakeExtensions = options;
-  }
-
-  protected handleConnection(conn: ConnType) {
-    if (this.getStatus() !== 'open') return;
-
-    this.log?.info(`new incoming connection`, {
-      ...conn.loggingMetadata,
-      clientId: this.clientId,
-    });
-
-    let session: Session<ConnType> | undefined = undefined;
-    const client = () => session?.to ?? 'unknown';
-
-    // kill the conn after the grace period if we haven't received a handshake
-    const handshakeTimeout = setTimeout(() => {
-      if (!session) {
-        this.log?.warn(
-          `connection to ${client()} timed out waiting for handshake, closing`,
-          {
-            ...conn.loggingMetadata,
-            clientId: this.clientId,
-            connectedTo: client(),
-          },
-        );
-        conn.telemetry?.span.setStatus({
-          code: SpanStatusCode.ERROR,
-          message: 'handshake timeout',
-        });
-        conn.close();
-      }
-    }, this.options.sessionDisconnectGraceMs);
-
-    const buffer: Array<Uint8Array> = [];
-    let receivedHandshakeMessage = false;
-
-    const handshakeHandler = (data: Uint8Array) => {
-      // if we've already received, just buffer the data
-      if (receivedHandshakeMessage) {
-        buffer.push(data);
-        return;
-      }
-
-      receivedHandshakeMessage = true;
-      clearTimeout(handshakeTimeout);
-
-      void this.receiveHandshakeRequestMessage(data, conn).then(
-        (maybeSession) => {
-          if (!maybeSession) {
-            conn.close();
-            return;
-          }
-
-          session = maybeSession;
-
-          // when we are done handshake sequence,
-          // remove handshake listener and use the normal message listener
-          const dataHandler = (data: Uint8Array) => {
-            const parsed = this.parseMsg(data, conn);
-            if (!parsed) {
-              conn.close();
-              return;
-            }
-
-            this.handleMsg(parsed, conn);
-          };
-
-          // process any data we missed
-          for (const data of buffer) {
-            dataHandler(data);
-          }
-
-          conn.removeDataListener(handshakeHandler);
-          conn.addDataListener(dataHandler);
-          buffer.length = 0;
-        },
-      );
-    };
-
-    conn.addDataListener(handshakeHandler);
-    conn.addCloseListener(() => {
-      if (!session) return;
-      this.log?.info(`connection to ${client()} disconnected`, {
-        ...conn.loggingMetadata,
-        clientId: this.clientId,
-      });
-      this.onDisconnect(conn, session);
-    });
-
-    conn.addErrorListener((err) => {
-      conn.telemetry?.span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: 'connection error',
-      });
-      if (!session) return;
-      this.log?.warn(
-        `connection to ${client()} got an error: ${coerceErrorString(err)}`,
-        { ...conn.loggingMetadata, clientId: this.clientId },
-      );
-    });
-  }
-
-  private async validateHandshakeMetadata(
-    conn: ConnType,
-    session: Session<ConnType> | undefined,
-    rawMetadata: Static<
-      typeof ControlMessageHandshakeRequestSchema
-    >['metadata'],
-    from: TransportClientId,
-  ): Promise<ParsedMetadata | false> {
-    let parsedMetadata: ParsedMetadata = {};
-    if (this.handshakeExtensions) {
-      // check that the metadata that was sent is the correct shape
-      if (!Value.Check(this.handshakeExtensions.schema, rawMetadata)) {
-        conn.telemetry?.span.setStatus({
-          code: SpanStatusCode.ERROR,
-          message: 'malformed handshake meta',
-        });
-        const reason = 'received malformed handshake metadata';
-        const responseMsg = handshakeResponseMessage(this.clientId, from, {
-          ok: false,
-          reason,
-        });
-        conn.send(this.codec.toBuffer(responseMsg));
-        this.log?.warn(`received malformed handshake metadata from ${from}`, {
-          ...conn.loggingMetadata,
-          clientId: this.clientId,
-          validationErrors: [
-            ...Value.Errors(this.handshakeExtensions.schema, rawMetadata),
-          ],
-        });
-        this.protocolError(ProtocolError.HandshakeFailed, reason);
-        return false;
-      }
-
-      const previousParsedMetadata = session
-        ? this.sessionHandshakeMetadata.get(session)
-        : undefined;
-
-      parsedMetadata = await this.handshakeExtensions.validate(
-        rawMetadata,
-        previousParsedMetadata,
-      );
-
-      // handler rejected the connection
-      if (parsedMetadata === false) {
-        const reason = 'rejected by handshake handler';
-        conn.telemetry?.span.setStatus({
-          code: SpanStatusCode.ERROR,
-          message: reason,
-        });
-        const responseMsg = handshakeResponseMessage(this.clientId, from, {
-          ok: false,
-          reason,
-        });
-        conn.send(this.codec.toBuffer(responseMsg));
-        this.log?.warn(`rejected handshake from ${from}`, {
-          ...conn.loggingMetadata,
-          clientId: this.clientId,
-        });
-        this.protocolError(ProtocolError.HandshakeFailed, reason);
-        return false;
-      }
-    }
-
-    return parsedMetadata;
-  }
-
-  async receiveHandshakeRequestMessage(
-    data: Uint8Array,
-    conn: ConnType,
-  ): Promise<Session<ConnType> | false> {
-    const parsed = this.parseMsg(data, conn);
-    if (!parsed) {
-      conn.telemetry?.span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: 'non-transport message',
-      });
-      this.protocolError(
-        ProtocolError.HandshakeFailed,
-        'received non-transport message',
-      );
-      return false;
-    }
-
-    if (!Value.Check(ControlMessageHandshakeRequestSchema, parsed.payload)) {
-      conn.telemetry?.span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: 'invalid handshake request',
-      });
-      const reason = 'received invalid handshake msg';
-      const responseMsg = handshakeResponseMessage(this.clientId, parsed.from, {
-        ok: false,
-        reason,
-      });
-      conn.send(this.codec.toBuffer(responseMsg));
-      this.log?.warn(reason, {
-        ...conn.loggingMetadata,
-        clientId: this.clientId,
-        // safe to this.log metadata here as we remove the payload
-        // before passing it to user-land
-        transportMessage: parsed,
-        validationErrors: [
-          ...Value.Errors(ControlMessageHandshakeRequestSchema, parsed.payload),
-        ],
-      });
-      this.protocolError(
-        ProtocolError.HandshakeFailed,
-        'invalid handshake request',
-      );
-      return false;
-    }
-
-    // double check protocol version here
-    const gotVersion = parsed.payload.protocolVersion;
-    if (gotVersion !== PROTOCOL_VERSION) {
-      conn.telemetry?.span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: 'incorrect protocol version',
-      });
-
-      const reason = `incorrect version (got: ${gotVersion} wanted ${PROTOCOL_VERSION})`;
-      const responseMsg = handshakeResponseMessage(this.clientId, parsed.from, {
-        ok: false,
-        reason,
-      });
-      conn.send(this.codec.toBuffer(responseMsg));
-      this.log?.warn(
-        `received handshake msg with incompatible protocol version (got: ${gotVersion}, expected: ${PROTOCOL_VERSION})`,
-        { ...conn.loggingMetadata, clientId: this.clientId },
-      );
-      this.protocolError(ProtocolError.HandshakeFailed, reason);
-      return false;
-    }
-
-    const oldSession = this.sessions.get(parsed.from);
-    const parsedMetadata = await this.validateHandshakeMetadata(
-      conn,
-      oldSession,
-      parsed.payload.metadata,
-      parsed.from,
-    );
-
-    if (parsedMetadata === false) {
-      return false;
-    }
-
-    const { session, isTransparentReconnect } = this.getOrCreateSession({
-      to: parsed.from,
-      conn,
-      sessionId: parsed.payload.sessionId,
-      propagationCtx: parsed.tracing,
-    });
-
-    this.sessionHandshakeMetadata.set(session, parsedMetadata);
-
-    this.log?.debug(
-      `handshake from ${parsed.from} ok, responding with handshake success`,
-      conn.loggingMetadata,
-    );
-    const responseMsg = handshakeResponseMessage(this.clientId, parsed.from, {
-      ok: true,
-      sessionId: session.id,
-    });
-    conn.send(this.codec.toBuffer(responseMsg));
-    this.onConnect(conn, session, isTransparentReconnect);
-
-    return session;
   }
 }
